requirements.txt
Flask==2.2.2
geopy==2.3.0
gurobipy==9.5.2
kafka-python==2.0.2
lkh==1.1.1
numpy==1.24.3
ortools==9.4.1874
pandas==1.5.1
requests==2.28.1
thrift==0.16.0
thriftpy2==0.4.16
tornado==6.2
tsplib95==0.7.1
matplotlib==3.5.3
future==0.18.3
termcolor==2.3.0
Werkzeug==2.2.2
tqdm==4.65.0
DingtalkChatbot==1.5.7
statsmodels==0.14.1
pmdarima==2.0.4
scikit-learn==1.2.2